Ajustes para relación User con Role medic y Specialty

+++ routes/web.php
@@ -4,6 +4,7 @@ use App\Http\Controllers\AppointmentController;
 use App\Http\Controllers\RoleController;
 use App\Http\Controllers\UserController;
 use App\Http\Controllers\SpecialtyController;
+use App\Http\Controllers\EmployeeController;
 use App\Livewire\EmployeeManagement;
 use App\Livewire\PatientManagement;
 use App\Livewire\UserManagement;
@@ -29,10 +30,10 @@ Route::middleware(['auth:sanctum', 'verified'])->group(function () {
     Route::get('/admin/patients', PatientManagement::class)->name('patients');
     //Route::get('employees', EmployeeManagement::class)->name('employees');
 
-    Route::get('/employees', EmployeeManagement::class)->name('employees.index');
-    Route::post('/employees', [UserController::class, 'store'])->name('employees.store');
-    Route::patch('/employees/{employee}', [UserController::class, 'update'])->name('employees.update');
-    Route::delete('/employees/{employee}', [UserController::class, 'destroy'])->name('employees.destroy');
+    Route::get('admin/employees', EmployeeManagement::class)->name('employees.index');
+    Route::post('admin//employees', [UserController::class, 'store'])->name('employees.store');
+    Route::patch('admin/employees/{employee}', [UserController::class, 'update'])->name('employees.update');
+    Route::delete('admin/employees/{employee}', [UserController::class, 'destroy'])->name('employees.destroy');
 });
 
 // Rutas protegidas para roles específicos
@@ -52,6 +53,9 @@ Route::middleware(['auth', 'role:admin|super-admin'])->group(function () {
         'update' => 'admin.appointments.update',
         'destroy' => 'admin.appointments.destroy',
     ]);
+    //Route::resource('admin/employees', EmployeeController::class);
+    //Route::get('admin/employees/assign-specialties/{id}', [EmployeeController::class, 'assignSpecialties'])->name('employees.assign.specialties');
+    //Route::post('admin/employees/store-specialties/{id}', [EmployeeController::class, 'storeSpecialties'])->name('employees.store.specialties');
     Route::get('medic/appointments', [AppointmentController::class, 'medicIndex'])->name('medic.appointments.index');
     Route::resource('specialties', SpecialtyController::class);
 });
